requirements/base.txt
djangorestframework~=3.12.4
Django~=3.2.7
django-filter~=21.1